Added permission to manifest for internet Created threads to make internet requests
